Show number of species in status bar and statistics
The number of species is the number of distinct `cladeID` in the organisms genetic code. I think it’s not really the same as species (it does not count for color changes), but it’s a good enough indicator for now.

Also refactored how the population/deaths/births are scaled on the chart: now they are using the same “max” value so we can tell what portion of the population dies