Synchronize all db queries
This should eliminate all exceptions that the auto analyzer introduced. I had to restructure the code so that the query execution and the ResultSet processing are both synchronized in the same block. I got lots of exceptions if I only synchronized the query itself and not the ResultSet processing, I believe because internally I only use one Statement, and the resources in the ResultSet gets destroyed or overwritten by the next query. Synchronization seems to solve the problem.