Optimize collision detection
It increased the frame rate from 35 to 43 on my machine (with rendering off).

The old version gathered all organisms that potentially overlapped with the given organism, and then put all organisms into a HashSet. In benchmarks this gathering consumed about 30% of the total cpu time.

The new version finds the buckets that overlaps with the given organisms, and simply iterates through the organisms in all those buckets, and stops at the one that actually has a collision. In benchmarks this new method “only” consumed about 15% of the total cpu time.

Note that the old matcher searched the colliding organisms in a pseudo random order (because all were put in a Set, and Sets have indeterministic order) while this new algorithm finds the first colliding organism in order. So the simulation outcome might differ.