Add scrollbar for “remarkable organisms”
I had to refactor a lot of code to make this possible. The content of the statistics window was destroyed and recreated every time the statistics were updated. This means the scrollbar reset its position. So I changed to create the UI components only once and then we just update the values.