Source from MarcoDBAA's announcement at 2023-08-30 21:23 HKT
Release notes

First: DEFAULT your parameters, because I changed default world area size to 7200 because of PLANKTON helping higher consumer symmetries. Detritus to CO2 divisor was increased to 900, because PLANKTON lost low maintenance. Special exceptions for PLANKTON-GRAY vs SKY and VIOLET were removed for now. BLACK 3/4 force EYEs without VISION to react on them (not the last segment seen), and it is invulnerable from "shown in color" frame 6 to 0 in general. An attack on it creates vulnerability from 10 counting down to 7. CREAM vs LAVENDER is now weakening the shield for 180 + CREAM segment length (instead of 179) in a frame. CREAM segments shorter than 5 are therefore less effective than in the old February version, above they are better