Only draw the organisms that are visible
This should give a few extra fps by skipping the redraw for organisms that are outside of the visible part of the screen. If your world is so small that it fits on the screen then this will make it a bit *slower*, but for big worlds this should add a few fps when you look at the world.