Use GroupLayout and splitpane in statistics window
GroupLayout seems to behave a bit better for small screen sizes than GridBagLayout, so we can have a scrollpane for the remarkable organisms panel and also for the right panel. Previously I was only able to add a scrollpane for the whole window, which did not look (and behave) amazing.