Auto balance load among threads
This should result in faster speeds because after every iteration fastest thread gets one more column, and one column is taken from the slowest thread. We might only want to do this rebalancing occasionally, not after each iteration to reduce overhead, but so far it seems to work.
The world I’m testing it with has almost all organisms in the left third, so this helps a little bit, although not as much as I expected (with 4 threads the speed goes up from 15 FPS to 20 FPS).