Source from MarcoDBAA's announcement at 2023-09-18 08:36 HKT
Release notes

Mirror is now part of the clade string, but mirror mutations for 1-symmetric organisms are ignored (they don´t do anything for them). If an organisms was mirrored a "=" is added before the hexadecimal clade part, if it was unmirrored a "#" is added before it. Removed the check for O2, when substances are released into the atmosphere. This should improve performance a bit (not sure if noticeable yet), and after killing them all, O2 could become minimally negative, but all organisms will disperse completely. If energy was transfered 100% correctly it would still be at 0.0 O2, but some rounding errors with (maybe irrational numbers) are possible. PLANKTON with killers, infectious and protective effectivity is reduced from 0.82 to 0.8125, and VIOLET-PLANKTON is added to that group. C4-only was buffed a little bit (hopefully not too much xD) and all BLOND-SPORE-7-12 organisms cannot reduce reproduction energy to less than 40 to minimize blue spore spam