Source from MarcoDBAA's announcement at 2023-12-24 06:53 HKT
Release notes

Default parameters for a slight CRIMSON nerf and CO turns faster into CO2 (nerf for C4). But C4 dodgers are more effective (same as Octobre version) at photosynthesis and maximum cap for C4 only photosynthesis was also increased. PLANKTON threshold for spinners and 1-symmetric ones was increased to 475 (nerf), but PLANKTON spinning filtrates (buff) a bit better, low maintenance was a bit improved, but not to the old value (still helping lower symmetries, but less than in the last version)