Source from MarcoDBAA's announcement on 2024-03-01
Release notes

DEFAULT your parameters, because the CO -> CO2 divisor was increased to 10000, and ROSE symbiont effectivity was a bit reduced. Buffed C4 dodgers effectivity again (but still not exactly as high as before the bug fix). And buffed C4-YELLOW to the value before the fix. ROSE-BLOND minimum reproduction energy is now at 21 for all of them, not only with WHITE, for balance and performance reasons. CORAL-YELLOW (with no plant/consumer segments) can now affect symbionts, if they have AUBURN, and also affects every victim for a bit longer in general. Removed the VIOLET nerf for viruses completely.